--- streampark-console/streampark-console-registry/streampark-registry-api/src/main/java/org/apache/streampark/registry/api/ha/DefaultServerStatusChangeListener.java
@@ -1,34 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.streampark.registry.api.ha;
-
-import lombok.extern.slf4j.Slf4j;
-
-@Slf4j
-public class DefaultServerStatusChangeListener extends AbstractServerStatusChangeListener {
-
-    @Override
-    public void changeToActive() {
-        log.info("The status is active now.");
-    }
-
-    @Override
-    public void changeToStandBy() {
-        log.info("The status is standby now.");
-    }
-}
