+++ streampark-console/streampark-console-service/src/main/java/org/apache/streampark/console/core/task/ConsoleHeartBeatTask.java
@@ -17,15 +17,15 @@
 
 package org.apache.streampark.console.core.task;
 
-import org.apache.streampark.common.enums.ServerStatus;
-import org.apache.streampark.common.lifecycle.ServerLifeCycleManager;
-import org.apache.streampark.common.model.BaseHeartBeatTask;
-import org.apache.streampark.common.model.ConsoleHeartBeat;
 import org.apache.streampark.common.utils.JSONUtils;
 import org.apache.streampark.common.utils.NetworkUtils;
 import org.apache.streampark.common.utils.OSUtils;
 import org.apache.streampark.console.core.config.ConsoleConfig;
 import org.apache.streampark.registry.api.RegistryClient;
+import org.apache.streampark.registry.api.enums.ServerStatusEnum;
+import org.apache.streampark.registry.api.lifecycle.ServerLifeCycleManager;
+import org.apache.streampark.registry.api.model.BaseHeartBeatTask;
+import org.apache.streampark.registry.api.model.ConsoleHeartBeat;
 
 import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
@@ -56,7 +56,7 @@ public class ConsoleHeartBeatTask extends BaseHeartBeatTask<ConsoleHeartBeat> {
             .startupTime(ServerLifeCycleManager.getServerStartupTime())
             .reportTime(System.currentTimeMillis())
             .processId(processId)
-            .serverStatus(ServerStatus.NORMAL)
+            .serverStatusEnum(ServerStatusEnum.NORMAL)
             .host(NetworkUtils.getHost())
             .port(consoleConfig.getListenPort())
             .build();
